# Implement wand-like optimization for boolean query