--- a/core/analysis/token_attributes.hpp
+++ b/core/analysis/token_attributes.hpp
@@ -162,26 +162,4 @@ class attribute_provider_change final : public attribute {
   mutable callback_f callback_{&noop};
 };
 
-// min score set by document consumers
-// *_max set by document producers
-struct score_threshold final : public attribute {
-  // DO NOT CHANGE NAME
-  static constexpr std::string_view type_name() noexcept {
-    return "iresearch::score_threshold";
-  }
-
-  score_t min;
-  // For disjunction/conjunction it's just sum of sub-iterators max score
-  // For iterator without score it depends on count of documents in iterator
-  // For wanderator it's max score for whole skip-list
-  // TODO(MBkkt) tail_max better here and not affect correctness
-  //  but to support it we need to know max value in the tail blocks.
-  //  Open question: how do it without read next blocks?
-  score_t list_max{};
-  const score_t* leaf_max{};
-#ifdef IRESEARCH_TEST                   // Used for tests
-  std::span<const score_t> levels_max;  // levels_max.back() == *leaf_max
-#endif
-};
-
 }  // namespace irs