--- a/core/formats/wand_writer.hpp
+++ b/core/formats/wand_writer.hpp
@@ -119,7 +119,7 @@ struct WandScorer<true> {
                const attribute_provider& attrs) {
     func_ =
       scorer_.prepare_scorer(reader, features, stats_.data(), attrs, kNoBoost);
-    return static_cast<bool>(func_);
+    return !func_.IsDefault();
   }
 
   score_t GetScore() const noexcept {
@@ -324,14 +324,18 @@ class FreqNormSource final : public WandSource {
 
   void Read(data_input& in, size_t size) final {
     freq_.value = in.read_vint();
+    // TODO(MBkkt) don't compute vsize here
+    const auto read = bytes_io<uint32_t>::vsize(freq_.value);
+    // We need to always try to read norm, because we have compatibility
+    // between BM25 in the index and TFIDF in the query
+    [[maybe_unused]] auto norm = freq_.value;
+    IRS_ASSERT(read <= size);
+    if (read != size) {
+      // TODO(MBkkt) if (!kNorm) in.skip(read - size);
+      norm += in.read_vint();
+    }
     if constexpr (kNorm) {
-      // TODO(MBkkt) don't compute vsize here
-      const auto read = bytes_io<uint32_t>::vsize(freq_.value);
-      norm_.value = freq_.value;
-      IRS_ASSERT(read <= size);
-      if (read != size) {
-        norm_.value += in.read_vint();
-      }
+      norm_.value = norm;
     }
   }