--- a/core/index/index_reader_options.hpp
+++ b/core/index/index_reader_options.hpp
@@ -47,6 +47,7 @@ inline constexpr size_t kMaxScorers = bits_required<uint64_t>();
 
 struct WandContext {
   static constexpr auto kDisable = std::numeric_limits<byte_type>::max();
+
   bool Enabled() const noexcept { return index != kDisable; }
 
   // Index of the wand data in the IndexWriter to use for optimization.