--- a/core/search/all_iterator.cpp
+++ b/core/search/all_iterator.cpp
@@ -27,15 +27,15 @@
 namespace irs {
 
 AllIterator::AllIterator(const SubReader& reader, const byte_type* query_stats,
-                         const Scorers& order, uint64_t docs_count, score_t boost)
+                         const Scorers& order, uint64_t docs_count,
+                         score_t boost)
   : max_doc_{doc_id_t(doc_limits::min() + docs_count - 1)} {
   std::get<cost>(attrs_).reset(max_doc_);
 
   if (!order.empty()) {
     auto& score = std::get<irs::score>(attrs_);
-    score =
-      CompileScore(order.buckets(), reader, irs::empty_term_reader(docs_count),
-                   query_stats, *this, boost);
+    CompileScore(score, order.buckets(), reader,
+                 irs::empty_term_reader(docs_count), query_stats, *this, boost);
   }
 }