… core/search/disjunction.hpp
… core/search/filter.hpp
… core/search/min_match_disjunction.hpp
… core/search/multiterm_query.cpp
… core/search/nested_filter.cpp
… core/search/ngram_similarity_filter.cpp
… core/search/ngram_similarity_query.cpp
… core/search/phrase_iterator.hpp
… core/search/phrase_query.cpp
… core/search/same_position_filter.cpp
… core/search/score.cpp
… core/search/score.hpp
… core/search/score_function.cpp
… core/search/score_function.hpp
… core/search/term_query.cpp
… core/search/tfidf.cpp
… core/utils/misc.hpp
… tests/formats/formats_15_tests.cpp
… tests/index/assert_format.cpp
… tests/search/all_filter_tests.cpp
… tests/search/boolean_filter_tests.cpp
… tests/search/filter_test_case_base.hpp
… tests/search/nested_filter_test.cpp
… tests/search/ngram_similarity_filter_tests.cpp
… tests/search/sort_tests.cpp
… tests/search/terms_filter_test.cpp
… tests/search/top_terms_collector_test.cpp
… tests/search/wand_test.cpp
… utils/index-put.cpp
… utils/index-search.cpp